Add custom GameInstance class
To avoid recalculating the fibonacci numbers and because this function doesnt belong
to Destructible piece I implemented a custom GameInstance class.

The reason to do this is because I want to have a kind of global access to the fibs but
also I want to calculate them once. Since GameInstance is unique and accesible from every actor
is a good place to keep this function.